Put some common testing code in one module
