+++ codeworld-prediction/codeworld-prediction.cabal
@@ -22,6 +22,7 @@ Test-Suite test-prediction
     Main-is:           prediction.hs
     Type:              exitcode-stdio-1.0
     Other-modules:     CodeWorld.Prediction
+                       Common
     Build-depends:     base,
                        containers,
                        QuickCheck
@@ -30,7 +31,8 @@ Test-Suite test-prediction-paper
     Hs-source-dirs:    tests, src
     Main-is:           prediction-paper.hs
     Type:              exitcode-stdio-1.0
-    Other-modules:     CodeWorld.Prediction
+    Other-modules:     CodeWorld.Prediction.Paper
+                       Common
     Build-depends:     base,
                        containers,
                        QuickCheck
