+++ codeworld-prediction/tests/prediction-paper.hs
@@ -25,8 +25,8 @@ import Data.List
 import System.Exit
 import Test.QuickCheck
 
-type TimeStamp = Double
-type TEvent = (TimeStamp, Maybe Event)
+import Common hiding (Event)
+
 
 type EventsTodo = (IM.IntMap [TEvent])
 type EventsDone = (IM.IntMap [TEvent])
@@ -36,8 +36,6 @@ type Book = [LogEntry]
 
 type Check = (EventsDone, (Int, TEvent))
 type CheckReport = (Check, Log Book, Log Book, Log Book)
-
--- Fake state and handle functions
 
 -- Fake state and handle functions
 
@@ -55,43 +53,6 @@ instance Game Book where
 rate :: Double
 rate = 1/4 -- decimal display
 
-
--- Generation of random schedules
-
-newtype GenEventsTodo = GenEventsTodo EventsTodo deriving Show
-
-genTimeStamps :: Gen [Double]
-genTimeStamps = fmap (drop 1 . scanl (+) 0) $
-    sized $ \n -> do
-        k <- choose (0,n)
-        sequence [ genOne | _ <- [1..k] ]
-  where
-    genOne = frequency [(5, getPositive <$> arbitrary), (1, return 0)]
-
-instance Arbitrary GenEventsTodo where
-    arbitrary = do
-        -- get ascending positive timestamps
-        tss <- genTimeStamps
-        qts <- sublistOf tss
-        p1ts  <- sublistOf tss
-        p2ts  <- sublistOf tss
-        -- some are just pings, some are real events
-        p1 <- traverse (makePingOrEvent 'A') p1ts
-        p2 <- traverse (makePingOrEvent 'B') p2ts
-        return $ GenEventsTodo (IM.fromList [(0,p1), (1,p2)])
-      where makePingOrEvent i ts = do
-               coin <- arbitrary
-               if coin then return $ (ts, Nothing)
-                       else return $ (ts, Just i)
-
-    -- shrinking removes poitns of the list, and
-    -- further reduces singleton lists mildly
-    -- (the original shrink would get us too close to epsilon)
-    shrink (GenEventsTodo m) =
-        map GenEventsTodo $
-        unShrinkOne $ traverse (listShrinkOne' (firstA mildlySmaller)) m
-      where mildlySmaller x | x > 1     = [x - 1]
-                            | otherwise = []
 
 -- The actual check:
 -- Exhaustively search the order in which these events could happen
@@ -138,10 +99,12 @@ failedChecks schedule = map mkReport $ filter (not . check) allChecks
 
 -- The quickcheck test, with reporting
 
-testPrediction :: Blind GenEventsTodo -> Property
-testPrediction (Blind (GenEventsTodo schedule)) = do
-   reportFailedCheck (head failed) `whenFail` null failed
-  where failed = failedChecks schedule
+testPrediction :: Property
+testPrediction = forAllSchedules $ \(_, schedule) ->
+    let failed = failedChecks schedule
+    in reportFailedCheck (head failed) `whenFail` null failed
+
+
 
 reportFailedCheck :: CheckReport -> IO ()
 reportFailedCheck (c, before, expected, actual) = do
@@ -175,42 +138,3 @@ main = do
         _          -> exitFailure
   where
     args = stdArgs { maxSize = 30 } -- more gets too large
-
--- QuickCheck utilities
-
-firstA :: Applicative f => (t -> f a1) -> (t, a) -> f (a1, a)
-firstA f (a,b) = (,) <$> f a <*> pure b
-
--- An applicative functor that shrinks exactly one element
--- Also see http://stackoverflow.com/a/41944525/946226
-data ShrinkOne a = ShrinkOne a [a]
-
-instance Functor ShrinkOne where
-    fmap f (ShrinkOne o s) = ShrinkOne (f o) (map f s)
-
-instance Applicative ShrinkOne where
-    pure x = ShrinkOne x []
-    ShrinkOne f fs <*> ShrinkOne x xs = ShrinkOne (f x) (map ($x) fs ++ map f xs)
-
-shrinkOne :: Arbitrary a => a -> ShrinkOne a
-shrinkOne x = ShrinkOne x (shrink x)
-
-unShrinkOne :: ShrinkOne t -> [t]
-unShrinkOne (ShrinkOne _ xs) = xs
-
--- Remove one element of the list
-listShrinkOne :: [a] -> ShrinkOne [a]
-listShrinkOne xs = ShrinkOne xs (listShrink xs)
-  where
-    listShrink []     = []
-    listShrink (y:ys) = [ ys ] ++ [ y:ys' | ys' <- listShrink ys ]
-
--- Remove one element of the list, or reduce a singleton with
--- the given function.
-listShrinkOne' :: (a -> [a]) -> [a] -> ShrinkOne [a]
-listShrinkOne' _ []  = ShrinkOne [] []
-listShrinkOne' f [x] = ShrinkOne [x] (map (:[]) (f x))
-listShrinkOne' _ xs  = listShrinkOne xs
-
-
-
