+++ codeworld-prediction/tests/prediction.hs
@@ -23,11 +23,8 @@ import Data.List
 import System.Exit
 import Test.QuickCheck
 
-type Event = Int
-type TimeStamp = Double
-type TEvent = (TimeStamp, Maybe Event)
+import Common
 
-type EventsTodo = ([TimeStamp], IM.IntMap [TEvent])
 type EventsDone = ([TimeStamp], IM.IntMap [TEvent])
 
 type LogEntry = Either Double Event
@@ -51,47 +48,12 @@ rate = 1/4 -- decimal display
 
 -- Generation of random schedules
 
-newtype GenEventsTodo = GenEventsTodo EventsTodo deriving Show
-
-genTimeStamps :: Gen [Double]
-genTimeStamps = fmap (drop 1 . scanl (+) 0) $
-    sized $ \n -> do
-        k <- choose (0,n)
-        sequence [ genOne | _ <- [1..k] ]
-  where
-    genOne = frequency [(5, getPositive <$> arbitrary), (1, return 0)]
-
-instance Arbitrary GenEventsTodo where
-    arbitrary = do
-        -- get ascending positive timestamps
-        tss <- genTimeStamps
-        qts <- sublistOf tss
-        p1ts  <- sublistOf tss
-        p2ts  <- sublistOf tss
-        -- some are just pings, some are real events
-        p1 <- traverse (makePingOrEvent 0) p1ts
-        p2 <- traverse (makePingOrEvent 1) p2ts
-        return $ GenEventsTodo (qts, IM.fromList [(0,p1), (1,p2)])
-      where makePingOrEvent i ts = do
-               coin <- arbitrary
-               if coin then return $ (ts, Nothing)
-                       else return $ (ts, Just i)
-
-    -- shrinking removes poitns of the list, and
-    -- further reduces singleton lists mildly
-    -- (the original shrink would get us too close to epsilon)
-    shrink (GenEventsTodo (tss,m)) = map GenEventsTodo $ unShrinkOne $
-        ((,) <$> listShrinkOne' mildlySmaller tss
-             <*> traverse (listShrinkOne' (firstA mildlySmaller)) m)
-      where mildlySmaller x | x > 1     = [x - 1]
-                            | otherwise = []
-
 -- The actual check:
 -- Exhaustively search the order in which these events could happen
 -- Memoize every initial segment
 -- Ensure that all possible ways reach the same conclusion.
 
-failedChecks :: EventsTodo -> [CheckReport]
+failedChecks :: EventSchedule -> [CheckReport]
 failedChecks schedule = map mkReport $ filter (not . check) allChecks
   where
     allDone :: [EventsDone]
@@ -143,10 +105,10 @@ failedChecks schedule = map mkReport $ filter (not . check) allChecks
 
 -- The quickcheck test, with reporting
 
-testPrediction :: Blind GenEventsTodo -> Property
-testPrediction (Blind (GenEventsTodo schedule)) = do
-   reportFailedCheck (head failed) `whenFail` null failed
-  where failed = failedChecks schedule
+testPrediction :: Property
+testPrediction = forAllSchedules $ \schedule ->
+    let failed = failedChecks schedule
+    in reportFailedCheck (head failed) `whenFail` null failed
 
 reportFailedCheck :: CheckReport -> IO ()
 reportFailedCheck (c, before, expected, actual) = do
@@ -187,42 +149,4 @@ main = do
         _          -> exitFailure
   where
     args = stdArgs { maxSize = 30 } -- more gets too large
-
--- QuickCheck utilities
-
-firstA :: Applicative f => (t -> f a1) -> (t, a) -> f (a1, a)
-firstA f (a,b) = (,) <$> f a <*> pure b
-
--- An applicative functor that shrinks exactly one element
--- Also see http://stackoverflow.com/a/41944525/946226
-data ShrinkOne a = ShrinkOne a [a]
-
-instance Functor ShrinkOne where
-    fmap f (ShrinkOne o s) = ShrinkOne (f o) (map f s)
-
-instance Applicative ShrinkOne where
-    pure x = ShrinkOne x []
-    ShrinkOne f fs <*> ShrinkOne x xs = ShrinkOne (f x) (map ($x) fs ++ map f xs)
-
-shrinkOne :: Arbitrary a => a -> ShrinkOne a
-shrinkOne x = ShrinkOne x (shrink x)
-
-unShrinkOne :: ShrinkOne t -> [t]
-unShrinkOne (ShrinkOne _ xs) = xs
-
--- Remove one element of the list
-listShrinkOne :: [a] -> ShrinkOne [a]
-listShrinkOne xs = ShrinkOne xs (listShrink xs)
-  where
-    listShrink []     = []
-    listShrink (y:ys) = [ ys ] ++ [ y:ys' | ys' <- listShrink ys ]
-
--- Remove one element of the list, or reduce a singleton with
--- the given function.
-listShrinkOne' :: (a -> [a]) -> [a] -> ShrinkOne [a]
-listShrinkOne' _ []  = ShrinkOne [] []
-listShrinkOne' f [x] = ShrinkOne [x] (map (:[]) (f x))
-listShrinkOne' _ xs  = listShrinkOne xs
-
-
